--- src/tests/mocks/handlers.ts
@@ -1,4 +1,0 @@
-// import { testHandler } from "@/tests/mocks/handlers/testHandler";
-import { loginHandler } from "./handlers/loginHandler";
-
-export const handlers = [...loginHandler];
