+++ src/tests/mocks/handlers/index.ts
@@ -1,0 +1,4 @@
+import { loginHandler } from "@/tests/mocks/handlers/loginHandler";
+import { signUpHandler } from "@/tests/mocks/handlers/signUpHandler";
+
+export const handlers = [...loginHandler, ...signUpHandler];
